+++ packages/pwastudio/test/feature/index.test.js
@@ -5,5 +5,16 @@ import configTests from "../../../../test/feature/config-tests";
 import { given, when, then } from "./index.gwt";
 
 describe(`Validate that 'pwastudio' works`, () => {
-	configTests({ given, when, then });
+	const defaultCode = `/* -- */
+a {
+    /* test */
+}
+`;
+	configTests({ given, when, then, defaultCode });
+
+	test(`Ensure config can be used with CSS Modules`, async () => {
+		await when.configIsUsedWithCSSModules();
+
+		then.resultsShouldContainNoError();
+	});
 });